auto-append SWAN to bluetooth name at start

+++ phone/src/main/java/interdroid/swan/sensors/impl/CloudtestSensor.java
@@ -96,7 +96,7 @@ public class CloudtestSensor extends AbstractSwanSensor {
 
                     i ^= 1;
 
-                    putValueTrimSize(valuePath, id, start, i);
+                    putValueTrimSize(valuePath, id, start, (double)i);
 
                 try {
                     Thread.sleep(configuration.getInt(DELAY,mDefaultConfiguration.getInt(DELAY)));
